fix(testing): 用例级超时 timeout_seconds 生效（问题 3）
- execute_task 新增 max_duration 可选参数，透传至 _check_timeout
- _check_timeout 优先使用用例级超时，None 时回退全局配置
- BatchTestRunner._run_single 在 timeout_seconds>0 时透传用例级超时
- TaskContext.is_timeout 参数类型 int 放宽为 float
- 新增 5 个单元测试（透传/回退/生效/互不干扰），445 passed

+++ src/mobile_automation/models/task.py
@@ -166,13 +166,13 @@ class TaskContext:
         """
         return self.status in (TaskStatus.COMPLETED, TaskStatus.FAILED, TaskStatus.ABORTED)
 
-    def is_timeout(self, max_duration_seconds: int = 300) -> bool:
+    def is_timeout(self, max_duration_seconds: float = 300) -> bool:
         """
         判断任务是否已超时。
 
         参数
         ----------
-        max_duration_seconds : int
+        max_duration_seconds : float
             最大允许持续时间（秒），默认 300 秒（5 分钟）。
 
         返回
